Refactor Flask routes into blueprints
