+++ config.py
@@ -1,0 +1,5 @@
+"""Application-wide configuration helpers and constants."""
+
+from __future__ import annotations
+
+# TODO: Populate with configuration loading utilities and shared constants.
