+++ db/__init__.py
@@ -1,0 +1,3 @@
+"""Package placeholder for future db functionality."""
+
+from __future__ import annotations
