+++ igdb/cache.py
@@ -1,0 +1,5 @@
+"""IGDB cache management utilities."""
+
+from __future__ import annotations
+
+# TODO: Implement caching strategies for IGDB metadata and asset downloads.
