+++ igdb/diff.py
@@ -1,0 +1,5 @@
+"""Utilities for building IGDB diff reports."""
+
+from __future__ import annotations
+
+# TODO: Implement diffing logic for comparing IGDB data against cached records.
