+++ ingestion/data_loader.py
@@ -1,0 +1,5 @@
+"""Data ingestion pipeline and processed-game maintenance utilities."""
+
+from __future__ import annotations
+
+# TODO: Build functions to load games, resequence data, and backfill metadata.
