+++ jobs/__init__.py
@@ -1,0 +1,3 @@
+"""Package placeholder for future jobs functionality."""
+
+from __future__ import annotations
