+++ jobs/manager.py
@@ -1,0 +1,5 @@
+"""Background job tracking and orchestration utilities."""
+
+from __future__ import annotations
+
+# TODO: Implement background job dataclasses and manager logic for async tasks.
