+++ lookups/config.py
@@ -1,0 +1,5 @@
+"""Lookup metadata configuration and normalization utilities."""
+
+from __future__ import annotations
+
+# TODO: Implement helpers for loading lookup table metadata and normalization rules.
