+++ lookups/service.py
@@ -1,0 +1,5 @@
+"""Lookup-table persistence and service-layer helpers."""
+
+from __future__ import annotations
+
+# TODO: Implement services for managing lookup records and related joins.
