+++ media/covers.py
@@ -1,0 +1,5 @@
+"""Cover image discovery, processing, and encoding utilities."""
+
+from __future__ import annotations
+
+# TODO: Implement helpers for manipulating and locating cover images.
