+++ processed/duplicates.py
@@ -1,0 +1,5 @@
+"""Duplicate detection and cleanup utilities for processed games."""
+
+from __future__ import annotations
+
+# TODO: Implement duplicate scanning, merge strategies, and cleanup workflows.
