+++ processed/navigator.py
@@ -1,0 +1,5 @@
+"""Navigation helpers for iterating through processed games."""
+
+from __future__ import annotations
+
+# TODO: Implement navigation state management and skip-queue handling.
