+++ processed/source_index_cache.py
@@ -1,0 +1,5 @@
+"""Utilities for caching and mapping processed source index data."""
+
+from __future__ import annotations
+
+# TODO: Add helpers to manage in-memory caches for processed source indices.
