+++ routes/web.py
@@ -1,0 +1,8 @@
+"""HTML-facing Flask routes and authentication flows."""
+
+from __future__ import annotations
+
+from flask import Blueprint
+
+# TODO: Register HTML routes, authentication handlers, and template context setup.
+web_blueprint = Blueprint("web", __name__)
