+++ services/summaries.py
@@ -1,0 +1,5 @@
+"""Auxiliary services for generating summaries and other derived content."""
+
+from __future__ import annotations
+
+# TODO: Implement integrations for summary generation services.
