+++ templates/index.html
@@ -19,8 +19,8 @@
             <a class="topbar-brand" href="{{ url_for('index') }}">Game Editor</a>
             <nav class="topbar-nav" aria-label="Global">
                 <a class="topbar-link is-active" href="{{ url_for('index') }}">Editor</a>
-                <a class="topbar-link" href="{{ url_for('updates_page') }}">IGDB Updates</a>
-                <a class="topbar-link" href="{{ url_for('lookups_page') }}">Lookups</a>
+                <a class="topbar-link" href="{{ url_for('updates.updates_page') }}">IGDB Updates</a>
+                <a class="topbar-link" href="{{ url_for('lookups.lookups_page') }}">Lookups</a>
                 <a class="topbar-link" href="{{ url_for('logout') }}">Logout</a>
             </nav>
         </div>
