+++ updates/service.py
@@ -1,0 +1,5 @@
+"""Update caching, refresh orchestration, and related services."""
+
+from __future__ import annotations
+
+# TODO: Implement services for assembling update lists and managing refresh jobs.
