+++ web/app_factory.py
@@ -1,0 +1,7 @@
+"""Flask application factory and service client initialization."""
+
+from __future__ import annotations
+
+from flask import Flask
+
+# TODO: Provide factory helpers to configure the Flask app and integrate services.
